WIP: DRY up pipeline functionality. #25
